reorg transformers a bit
