+++ lib/entities/request/ClassicRequest.ts
@@ -1,8 +1,8 @@
 import { Protocol } from '@uniswap/router-sdk';
 import { BigNumber } from 'ethers';
 
-import { QuoteRequest, QuoteRequestInfo, RoutingType } from '.';
 import { DUMMY_GAS_WEI } from '../../constants';
+import { QuoteRequest, QuoteRequestInfo, RoutingType } from '.';
 import { DutchLimitRequest } from './DutchLimitRequest';
 
 export interface ClassicConfig {
