+++ lib/providers/transformers/QuoteTransformers/OnlyConfiguredQuotersFilter.ts
@@ -1,7 +1,7 @@
 import Logger from 'bunyan';
 
-import { QuoteTransformer } from '..';
 import { Quote, QuoteRequest } from '../../../entities';
+import { QuoteTransformer } from '..';
 
 // filters out any quote responses that came from unconfigured quoters
 // sometimes we have to receive quotes even when not requested by the user
