+++ lib/providers/transformers/index.ts
@@ -1,9 +1,7 @@
 import { Quote, QuoteRequest } from '../../entities';
 
-export * from './CompoundTransformer';
-export * from './QuoteTransformers/OnlyConfiguredQuotersFilter';
-export * from './QuoteTransformers/SyntheticUniswapXTransformer';
-export * from './QuoteTransformers/UniswapXOrderSizeFilter';
+export * from './QuoteTransformers';
+export * from './RequestTransformers';
 
 export interface QuoteTransformer {
   transform(requests: QuoteRequest[], quotes: Quote[]): Promise<Quote[]>;
